@@ -5,12 +5,12 @@ plugins {
 
 android {
     namespace 'com.example.calculator'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.example.calculator"
-        minSdk 21
-        targetSdk 32
+        minSdk 26
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -30,6 +30,9 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    buildFeatures {
+        viewBinding true
+    }
 }
 
 dependencies {
@@ -41,5 +44,5 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    apply plugin: 'kotlin-android-extensions'
+//    apply plugin: 'kotlin-android-extensions'
 }
